Re-organize declaration-related classes
