+++ src/main/kotlin/sigma/semantics/Prelude.kt
@@ -18,7 +18,7 @@ data class Prelude(
             )
 
             val (definitionBlock, _) = DeclarationScope.looped { innerDeclarationScopeLooped ->
-                val definitionBlock = LocalDefinitionBlock.build(
+                val definitionBlock = LocalValueDefinitionBlock.build(
                     typeScope = BuiltinTypeScope,
                     outerDeclarationScope = innerDeclarationScopeLooped,
                     definitions = prelude.definitions,
